pyproject.toml
[project]
name = "ComputHorde-facilitator"
requires-python = "==3.11.*"
version = "0.1.0"
dependencies = [
    "Django~=4.2.4",
    "django-constance[database]==3.1.0",
    "django-cors-headers~=4.2.0",
    "django-environ~=0.11.2",
    "django-extensions==3.2.3",
    "django-probes==1.7.0",
    "django-debug-toolbar==4.1.0",
    "celery~=5.3.1",
    "flower~=2.0.0",
    "gunicorn==20.1.0",
    "psycopg2-binary~=2.9.7",
    "redis~=4.6.0",
    "sentry-sdk==1.3.0",
    "django-fingerprint-rt==0.0.2",
    "requests>=2.31.0",
    "ipython>=8.22.1",
    "pydantic~=2.7.3",
    "django-structlog[celery]>=7.1.0",
    "django-allauth>=0.61.1",
    "channels-redis>=4.2.0",
    "channels[daphne]>=4.0.0",
    "boto3>=1.34.54",
    "bittensor==7.0.1",
    "more-itertools>=10.2.0",
    "uvicorn[standard]>=0.22.0",
    "torch>=2.2.1",
    "django-crispy-forms>=2.1",
    "crispy-bootstrap5>=2024.2",
    "djangorestframework>=3.15.1",
    "django-filter>=24.2",
    "wandb>=0.16.6",
    "transformers>=4.41.0",
    "tenacity>=8.3.0",
    "httpx>=0.27.0",
]
description = "ComputeHorde app facilitating interaction between end users and Validators"
authors = [{name = "Backend Developers LTD"}]
readme = "README.md"
license = {text = "Custom"}

[tool.pdm]
distribution = false

[[tool.pdm.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cpu"
include_packages = ["torch"]
exclude_packages = ["*"]

[tool.pdm.dev-dependencies]
test = [
    "pytest",
    "pytest-django",
    "pytest-xdist",
    "ipdb",
    "freezegun",
    "pytest-asyncio>=0.23.5",
    "responses>=0.25.0",
]
lint = [
    "ruff",
]
type_check = [
    "django-stubs[compatible-mypy]",
    "djangorestframework-stubs[compatible-mypy]",
    "mypy",
    "types-freezegun",
    "types-python-dateutil",
    "types-requests",
]
security_check = [
    "bandit>=1.7.7",
]
dev = [
    "nox>=2024.3.2",
    "cruft>=2.15.0",
    "ipdb>=0.13.13",
]

[tool.ruff]
src = ["app/src"]
line-length = 120

[tool.ruff.lint]
# TODO add D
select = [
    "E", "F", "I", "UP",
    "TCH005",
]
# TODO: remove E501 once docstrings are formatted
ignore = [
    "D100", "D105", "D107", "D200", "D202", "D203", "D205", "D212", "D400", "D401", "D415",
    "D101", "D102","D103", "D104", # TODO remove once we have docstring for all public methods
    "E501", # TODO: remove E501 once docstrings are formatted
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"test/**" = ["D", "F403", "F405"]

[tool.codespell]
skip = "*.min.js,pdm.lock,*.aof"
ignore-words-list = "datas,acount"
